requirements.txt
langchain-groq==0.3.2
pydantic==2.10.6
python-dotenv==1.0.1
pypdf==3.17.2
